requirements.txt
pillow
numpy
matplotlib
scipy
opencv-python
